pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "microbiome-tool"
version = "0.1.0"
description = "A research prototype for microbiome sample comparison using ASV tables, DuckDB, and FAISS."
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "faiss-cpu",
    "duckdb",
    "click",
    "fastapi",
    "uvicorn",
    "transformers",
    "torch",
]

[project.scripts]
load-reference = "cli.load_reference:load_reference"
compare-samples = "cli.compare_samples:compare_samples"
